% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkRvsCpp.R
\docType{data}
\name{benchmarkResults}
\alias{benchmarkResults}
\title{Results from running a performance benchmark on a personal computer 
including the time for parameter transformation}
\format{A data.table}
\usage{
benchmarkResults
}
\description{
Results from running a performance benchmark on a personal computer 
including the time for parameter transformation
}
\keyword{datasets}
