% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkRvsCpp.R
\docType{data}
\name{benchmarkResultsNoTransform}
\alias{benchmarkResultsNoTransform}
\title{Results from running a performance benchmark on a personal computer 
excluding the time for parameter transformation}
\format{A data.table}
\usage{
benchmarkResultsNoTransform
}
\description{
Results from running a performance benchmark on a personal computer 
excluding the time for parameter transformation
}
\keyword{datasets}
