% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{PCMBaseCppIsADevRelease}
\alias{PCMBaseCppIsADevRelease}
\title{Check if the PCMBaseCpp version correpsonds to a dev release}
\usage{
PCMBaseCppIsADevRelease()
}
\value{
a logical
}
\description{
We define a dev release as having a sub-release, eg 0.9.15.5 is
one whereas 0.9.16 is not. The number of components in the version can be
changed through the argument numVersionComponents.
}
