% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic.R
\name{PCMParamGetFullVector}
\alias{PCMParamGetFullVector}
\title{Get a vector with all model parameters unrolled}
\usage{
PCMParamGetFullVector(model, ...)
}
\arguments{
\item{model}{a PCM model object}

\item{...}{passed to methods}
}
\value{
a numerical vector
}
\description{
Get a vector with all model parameters unrolled
}
\examples{
PCMParamGetFullVector(PCMBase::PCMBaseTestObjects$model.a.123)
}
