% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic.R
\name{PCMTreePreorderCpp}
\alias{PCMTreePreorderCpp}
\title{Fast preorder of the edges in a tree}
\usage{
PCMTreePreorderCpp(tree)
}
\arguments{
\item{tree}{a phylo object}
}
\value{
an integer vector containing indices of rows in \code{tree$edge} in
their preorder order.
}
\description{
Fast preorder of the edges in a tree
}
\examples{
PCMTreePreorderCpp(PCMBase::PCMBaseTestObjects$tree.a)

}
