% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{PCMBaseCppIsADevRelease}
\alias{PCMBaseCppIsADevRelease}
\title{Check if the PCMBaseCpp version correpsonds to a dev release}
\usage{
PCMBaseCppIsADevRelease()
}
\value{
a logical
}
\description{
This function is used during unit-testing, to disable some unit-
tests which run extremely long or are consistently failing on some systems.
}
